// coach-connect--frontend/src/app/dashboard/profile/interfaces/profile-response.interface.ts
export interface ProfileResponse {
    phoneNumber:            null;
    emailConfirmationToken: null;
    photoUrl:               null;
    bio:                    null;
    _id:                    string;
    firstName:              string;
    lastName:               string;
    email:                  string;
    isEmailConfirmed:       boolean;
    __v:                    number;
}